+++ src/elements/Order.php
@@ -51,6 +51,7 @@ use craft\helpers\StringHelper;
 use craft\helpers\Template;
 use craft\helpers\UrlHelper;
 use craft\i18n\Locale;
+use craft\models\Site;
 use DateTime;
 use Throwable;
 use yii\base\Exception;
@@ -129,6 +130,7 @@ use yii\log\Logger;
  * @property-read string $storedTotalDiscountAsCurrency
  * @property-read string $storedTotalTaxAsCurrency
  * @property-read string $storedTotalTaxIncludedAsCurrency
+ * @property-read Site|null $orderSite
  * @property null|array|Address $estimatedBillingAddress
  * @property float $totalDiscount
  * @property null|array|Address $estimatedShippingAddress
@@ -601,19 +603,32 @@ class Order extends Element
     public $orderStatusId;
 
     /**
-     * The current order status ID. This will be null if the order is not complete
-     * and is still a cart.
+     * The language the cart was created in.
      *
-     * @var int|null Order status ID
+     * @var string The language the order was made in.
      * ---
      * ```php
-     * echo $order->orderStatusId;
+     * echo $order->orderLanguage;
      * ```
      * ```twig
-     * {{ order.orderStatusId }}
+     * {{ order.orderLanguage }}
      * ```
      */
     public $orderLanguage;
+
+    /**
+     * The site the order was created in.
+     *
+     * @var int|null Order site ID
+     * ---
+     * ```php
+     * echo $order->orderSiteId;
+     * ```
+     * ```twig
+     * {{ order.orderSiteId }}
+     * ```
+     */
+    public $orderSiteId;
 
 
     /**
@@ -1064,6 +1079,10 @@ class Order extends Element
             $this->orderLanguage = Craft::$app->language;
         }
 
+        if ($this->orderSiteId === null) {
+            $this->orderSiteId = Craft::$app->getSites()->getHasCurrentSite() ? Craft::$app->getSites()->getCurrentSite()->id : Craft::$app->getSites()->getPrimarySite()->id;
+        }
+
         if ($this->currency === null) {
             $this->currency = Plugin::getInstance()->getPaymentCurrencies()->getPrimaryPaymentCurrencyIso();
         }
@@ -1349,6 +1368,7 @@ class Order extends Element
         $names[] = 'shippingAddress';
         $names[] = 'shippingMethod';
         $names[] = 'transactions';
+        $names[] = 'orderSite';
         return $names;
     }
 
@@ -1841,6 +1861,7 @@ class Order extends Element
         $orderRecord->currency = $this->currency;
         $orderRecord->lastIp = $this->lastIp;
         $orderRecord->orderLanguage = $this->orderLanguage;
+        $orderRecord->orderSiteId = $this->orderSiteId;
         $orderRecord->origin = $this->origin;
         $orderRecord->paymentCurrency = $this->paymentCurrency;
         $orderRecord->customerId = $this->customerId;
@@ -3030,6 +3051,21 @@ class Order extends Element
     public function getOrderStatus()
     {
         return Plugin::getInstance()->getOrderStatuses()->getOrderStatusById($this->orderStatusId);
+    }
+
+    /**
+     * Get the site for the order.
+     *
+     * @return Site|null
+     * @since 3.x
+     */
+    public function getOrderSite()
+    {
+        if (!$this->orderSiteId) {
+            return null;
+        }
+
+        return Craft::$app->getSites()->getSiteById($this->orderSiteId);
     }
 
     /**