+++ src/web/assets/commerceui/src/js/order/apps/OrderMeta.vue
@@ -33,6 +33,13 @@
                 <date-ordered-input :date="draft.order.dateOrdered" @update="updateDateOrderedInput"></date-ordered-input>
             </field>
 
+            <field v-if="order.isCompleted" :label="$options.filters.t('Order Site', 'commerce')">
+                <order-site
+                    :originalOrderSiteId="originalDraft.order.orderSiteId"
+                    :order="order"
+                    @updateOrder="updateOrder"></order-site>
+            </field>
+
             <div class="field" id="isCompleted-field"
                  v-if="!draft.order.isCompleted">
                 <div class="heading">
@@ -91,6 +98,12 @@
                     <span class="value">{{draft.order.dateOrdered.date}} {{draft.order.dateOrdered.time}}</span>
                 </div>
             </template>
+
+            <div class="data">
+                <h5 class="heading">{{"Order Site"|t('commerce')}}</h5>
+                <span class="value"
+                      v-if="draft.order.orderSite">{{draft.order.orderSite.name}} ({{draft.order.orderSite.language}})</span>
+            </div>
 
             <template v-if="draft.order.isCompleted">
                 <div class="data">
@@ -195,6 +208,7 @@
     import debounce from 'lodash.debounce'
     import {mapActions, mapGetters, mapState} from 'vuex'
     import OrderStatus from '../components/meta/OrderStatus'
+    import OrderSite from '../components/meta/OrderSite'
     import ShippingMethod from '../components/meta/ShippingMethod'
     import DateOrderedInput from '../components/meta/DateOrderedInput'
     import Field from '../../base/components/Field'
@@ -205,6 +219,7 @@
 
         components: {
             OrderStatus,
+            OrderSite,
             ShippingMethod,
             DateOrderedInput,
             Field,