+++ CHANGELOG.md
@@ -3,12 +3,15 @@
 ## Unreleased
 
 ### Added
+- Added the ability to track which site an order was place from.
+- Added a “Order Site” column to the Order index page.
 - Added the ability to change the default tax category right from the tax category index page. ([#1499](https://github.com/craftcms/commerce/issues/1499))
 - Added the ability to change the default tax category right from the Tax Categories index page. ([#1499](https://github.com/craftcms/commerce/issues/1499))
 - Added the ability to change the default shipping category right from the Shipping Categories index page.
 - Added the ability to change statuses right from the Shipping Methods index page.
 - Added an “ID” column to the Products index page. ([#1787](https://github.com/craftcms/commerce/issues/1787))
 - All front-end success/fail flash messages are now customizable by passing a hashed `successMessage`/`failMessage` param with the request. ([#1801](https://github.com/craftcms/commerce/issues/1801))
+- Added `craft\commerce\elements\Order::$orderSiteId`.
 - Added `craft\commerce\services\Webhooks::EVENT_BEFORE_PROCESS_WEBHOOK`. ([#1799](https://github.com/craftcms/commerce/issues/1799))
 - Added `craft\commerce\services\Webhooks::EVENT_AFTER_PROCESS_WEBHOOK`. ([#1799](https://github.com/craftcms/commerce/issues/1799))
 - Added `craft\commerce\services\Customers::EVENT_BEFORE_SAVE_CUSTOMER`.